fix settings files
